requirements.txt
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
Flask-WTF==1.2.1
WTForms==3.1.1
bcrypt==4.1.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.6.1
yadisk==3.1.0
pandas==2.1.4
matplotlib==3.8.2
numpy==1.24.3
pytest==7.4.3
webdavclient3==3.14.6
openpyxl==3.1.2
